update the dealbot consumer example and SDK for consumer
